refactor(projects): refactor cards component
