+++ src/components/projects-card/projects-card.style.ts
